fix login issue
