+++ .env.example
@@ -1,6 +1,8 @@
-# API server port
 PORT=8080
-# Network proxy server address
-PROXY=socks5://ip:port
-# Imitate access_token
+TZ=Asia/Shanghai
+PROXY=
+OPENAI_EMAIL=
+OPENAI_PASSWORD=
+CONTINUE_SIGNAL=
+ENABLE_HISTORY=
 IMITATE_ACCESS_TOKEN=
