+++ .gitignore
@@ -1,3 +1,4 @@
 .env
 go-chatgpt-api
 __debug*
+chat.openai.com.har
