+++ api/imitate/api.go
@@ -177,8 +177,8 @@ func sendConversationRequest(c *gin.Context, request chatgpt.CreateConversationR
 	req.Header.Set("User-Agent", api.UserAgent)
 	req.Header.Set(api.AuthorizationHeader, accessToken)
 	req.Header.Set("Accept", "text/event-stream")
-	if chatgpt.PUID != "" {
-		req.Header.Set("Cookie", "_puid="+chatgpt.PUID)
+	if api.PUID != "" {
+		req.Header.Set("Cookie", "_puid="+api.PUID)
 	}
 	resp, err := api.Client.Do(req)
 	if err != nil {
